+++ src/Detail/type/lecueBookType.ts
@@ -1,0 +1,6 @@
+export interface postedStickerType {
+  postedStickerId: number;
+  stickerImage: string;
+  positionX: number;
+  positionY: number;
+}
